Added new dockerfile and adjusted build scripts
