+++ bin/config.json
@@ -1,0 +1,1 @@
+{"Version":1,"InitialAurumSupply":500000000000000,"Port":"52000","BlockProductionInterval":"10s","Localhost":false}
