+++ bin/genesis_hashes.txt
@@ -1,0 +1,1 @@
+23aafe84f813bd5093599691ea5731425effe1b8c3f7c1e3c049012558160b8c
