--- cmd/config.json
@@ -1,1 +1,0 @@
-{"Version":1,"InitialAurumSupply":50000000,"Port":"5000","BlockProductionInterval":"10s","Localhost":false}
